requirements.txt
aenum==2.0.8
antlr4-python3-runtime==4.7.1
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.4.2
flake8==3.5.0
flake8-polyfill==1.0.2
idna==2.6
Jinja2==2.10
mando==0.6.4
MarkupSafe==1.0
mccabe==0.6.1
pip-autoremove==0.9.0
pycodestyle==2.3.1
pyflakes==1.6.0
radon==2.2.0
requests==2.18.4
rpy2==2.9.2
six==1.11.0
urllib3==1.22
